backend/pyproject.toml
[project]
name = "trendscope-backend"
version = "0.1.0"
description = "Stock trend analysis backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "yfinance>=0.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "fastapi>=0.100.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.23.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "statsmodels>=0.14.0",
    "TA-Lib>=0.4.0",
    "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.0.290",
    "mypy>=1.5.0",
    "black>=23.0.0",
    "httpx>=0.24.0",
]
ml = [
    "tensorflow>=2.13.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "C4", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/trendscope_backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
